chore(ios):  renames install/launch-app to install/launch-app-emulator
