+++ lib/targets/ios/target.js
@@ -1,11 +1,12 @@
 const CoreObject       = require('core-object');
 const path             = require('path');
 
-const bootEm           = require('./tasks/boot-emulator');
 const buildEm          = require('./tasks/build-emulator');
+
+const bootEm           = require('./tasks/boot-emulator');
 const openEm           = require('./tasks/open-emulator');
-const installApp       = require('./tasks/install-app');
-const launchApp        = require('./tasks/launch-app');
+const installAppEm     = require('./tasks/install-app-emulator');
+const launchAppEm      = require('./tasks/launch-app-emulator');
 
 const cdvConfig        = require('../cordova/utils/get-config');
 const cdvPath          = require('../cordova/utils/get-path');
@@ -49,9 +50,9 @@ module.exports = CoreObject.extend({
   run() {
     let emulatorId = this.device.uuid;
 
-    return bootEm(this.device)
+    return bootEm(emulatorId)
       .then(() => openEm())
-      .then(() => installApp(emulatorId, this.builtPath))
-      .then(() => launchApp(emulatorId, this.packageName))
+      .then(() => installAppEm(emulatorId, this.builtPath))
+      .then(() => launchAppEm(emulatorId, this.packageName))
   }
 });
