+++ lib/targets/ios/tasks/install-app-emulator.js
