+++ lib/targets/ios/tasks/launch-app-emulator.js
