+++ node-tests/unit/targets/ios/tasks/launch-app-test-emulator.js
@@ -3,27 +3,27 @@ const expect     = require('../../../../helpers/expect');
 const Promise    = require('rsvp').Promise;
 
 const emulatorId = 'emulatorId';
-const ipaPath    = 'ipa-path';
+const appName    = 'appName';
 
 const spawnArgs = [
   '/usr/bin/xcrun',
   [
     'simctl',
-    'install',
+    'launch',
     emulatorId,
-    ipaPath
+    appName
   ]
 ];
 
-describe('IOS Install App', function() {
-  let installApp;
+describe('IOS Launch App Emulator', () => {
+  let launchApp;
   let spawn;
 
   beforeEach(() => {
     spawn = td.replace('../../../../../lib/utils/spawn');
     td.when(spawn(...spawnArgs)).thenReturn(Promise.resolve({ code: 0 }));
 
-    installApp = require('../../../../../lib/targets/ios/tasks/install-app');
+    launchApp = require('../../../../../lib/targets/ios/tasks/launch-app-emulator');
   });
 
   afterEach(() => {
@@ -36,22 +36,22 @@ describe('IOS Install App', function() {
     td.when(spawn(), { ignoreExtraArgs: true })
       .thenReturn(Promise.resolve());
 
-    return installApp(emulatorId, ipaPath).then(() => {
+    return launchApp(emulatorId, appName).then(() => {
       td.verify(spawn(...spawnArgs));
 
       td.config({ ignoreWarnings: false });
     });
   });
 
-  it('spawns scrun and resolves with exit code', () => {
-    return expect(installApp(emulatorId, ipaPath))
+  it('spawns xcrun and resolves with exit code', () => {
+    return expect(launchApp(emulatorId, appName))
       .to.eventually.deep.equal({ code: 0 });
   });
 
   it('bubbles up error message when spawn rejects', () => {
     td.when(spawn(...spawnArgs)).thenReturn(Promise.reject('spawn error'));
 
-    return expect(installApp(emulatorId, ipaPath))
+    return expect(launchApp(emulatorId, appName))
       .to.eventually.be.rejectedWith('spawn error');
   });
 });
